[Gulp] Split Gulp file into Admin and Shop bundles
