+++ src/Sylius/Bundle/AdminBundle/package.json
@@ -5,9 +5,7 @@
   },
   "devDependencies": {
     "gulp": "^3.9.0",
-    "gulp-chug": "^0.5",
     "gulp-concat": "^2.6.0",
-    "gulp-debug": "^2.1.2",
     "gulp-if": "^2.0.0",
     "gulp-livereload": "^3.8.1",
     "gulp-order": "^1.1.1",
@@ -22,7 +20,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/Sylius/Sylius.git"
+    "url": "git+https://github.com/Sylius/SyliusAdminBundle.git"
   },
   "author": "Paweł Jędrzejewski",
   "license": "MIT"
