+++ src/Sylius/Bundle/ShopBundle/package.json
@@ -5,7 +5,6 @@
   },
   "devDependencies": {
     "gulp": "^3.9.0",
-    "gulp-chug": "^0.5",
     "gulp-concat": "^2.6.0",
     "gulp-debug": "^2.1.2",
     "gulp-if": "^2.0.0",
@@ -18,7 +17,7 @@
     "merge-stream": "^1.0.0"
   },
   "scripts": {
-    "gulp": "gulp"
+      "gulp": "gulp"
   },
   "repository": {
     "type": "git",
